version 0.23.123

+++ Terminator/Data/UserDataMain_Stage.cs
@@ -77,7 +77,7 @@ public partial class UserDataMain
                             continue;
                     }
 
-                    destination.value = (byte)source.conditionValue;
+                    destination.value = source.conditionValue;
                     
                     if(results == null)
                         results = new  List<LevelQuest>();
